% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{glpk_get_col_prim}
\alias{glpk_get_col_prim}
\title{Get the primal value of a variable}
\usage{
glpk_get_col_prim(optimizer, variable)
}
\arguments{
\item{optimizer}{the GLPK optimizer}

\item{variable}{a scalar affine term}
}
\description{
Get the primal value of a variable
}
