% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\docType{class}
\name{GLPK_optimizer-class}
\alias{GLPK_optimizer-class}
\alias{moi_add_variable,GLPK_optimizer-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_scalar_affine_function,MOI_less_than_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_scalar_affine_function,MOI_equal_to_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_scalar_affine_function,MOI_greater_than_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_scalar_affine_term,MOI_greater_than_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_scalar_affine_term,MOI_less_than_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_scalar_affine_term,MOI_equal_to_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_single_variable,MOI_greater_than_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_single_variable,MOI_integer_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_single_variable,MOI_zero_one_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_single_variable,MOI_interval_set-method}
\alias{moi_add_constraint,GLPK_optimizer,MOI_single_variable,MOI_less_than_set-method}
\alias{moi_set,GLPK_optimizer,MOI_objective_function,MOI_scalar_affine_function,missing-method}
\alias{moi_set,GLPK_optimizer,MOI_objective_function,MOI_scalar_affine_term,missing-method}
\alias{moi_set,GLPK_optimizer,MOI_objective_function,numeric,missing-method}
\alias{moi_set,GLPK_optimizer,MOI_objective_sense,MOI_optimization_sense,missing-method}
\alias{moi_optimize,GLPK_optimizer-method}
\alias{moi_get,GLPK_optimizer,MOI_variable_primal_attribute,MOI_variable_index-method}
\alias{moi_get,GLPK_optimizer,MOI_termination_status,missing-method}
\alias{moi_get,GLPK_optimizer,MOI_number_of_variables,missing-method}
\alias{moi_get,GLPK_optimizer,MOI_number_of_constraints,missing-method}
\alias{moi_get,GLPK_optimizer,MOI_objective_value,missing-method}
\alias{moi_get,GLPK_optimizer,MOI_primal_status,missing-method}
\alias{moi_get,GLPK_optimizer,MOI_result_count,missing-method}
\title{A MOI compatible GLPK optimizer}
\usage{
\S4method{moi_add_variable}{GLPK_optimizer}(model)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_scalar_affine_function,MOI_less_than_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_scalar_affine_function,MOI_equal_to_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_scalar_affine_function,MOI_greater_than_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_scalar_affine_term,MOI_greater_than_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_scalar_affine_term,MOI_less_than_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_scalar_affine_term,MOI_equal_to_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_single_variable,MOI_greater_than_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_single_variable,MOI_integer_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_single_variable,MOI_zero_one_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_single_variable,MOI_interval_set}(model, func, set)

\S4method{moi_add_constraint}{GLPK_optimizer,MOI_single_variable,MOI_less_than_set}(model, func, set)

\S4method{moi_set}{GLPK_optimizer,MOI_objective_function,MOI_scalar_affine_function,missing}(model, type, index, value)

\S4method{moi_set}{GLPK_optimizer,MOI_objective_function,MOI_scalar_affine_term,missing}(model, type, index, value)

\S4method{moi_set}{GLPK_optimizer,MOI_objective_function,numeric,missing}(model, type, index, value)

\S4method{moi_set}{GLPK_optimizer,MOI_objective_sense,MOI_optimization_sense,missing}(model, type, index, value)

\S4method{moi_optimize}{GLPK_optimizer}(model)

\S4method{moi_get}{GLPK_optimizer,MOI_variable_primal_attribute,MOI_variable_index}(model, type, index)

\S4method{moi_get}{GLPK_optimizer,MOI_termination_status,missing}(model, type)

\S4method{moi_get}{GLPK_optimizer,MOI_number_of_variables,missing}(model, type)

\S4method{moi_get}{GLPK_optimizer,MOI_number_of_constraints,missing}(model, type)

\S4method{moi_get}{GLPK_optimizer,MOI_objective_value,missing}(model, type)

\S4method{moi_get}{GLPK_optimizer,MOI_primal_status,missing}(model, type)

\S4method{moi_get}{GLPK_optimizer,MOI_result_count,missing}(model, type)
}
\description{
A MOI compatible GLPK optimizer
}
\keyword{internal}
