% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{GLPK_optimizer}
\alias{GLPK_optimizer}
\title{A GLPK Optimizer}
\usage{
GLPK_optimizer(presolve = FALSE)
}
\arguments{
\item{presolve}{a length 1 logical if presolve should be used.}
}
\description{
A GLPK Optimizer
}
