% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{set_irowgen_callback}
\alias{set_irowgen_callback}
\title{Set an irowgen callback}
\usage{
set_irowgen_callback(optimizer, fun)
}
\arguments{
\item{optimizer}{a GLPK optimizer}

\item{fun}{the function that should be called when an integer solution has
been found.}
}
\description{
Set an irowgen callback
}
