% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{GLPK_optimizer}
\alias{GLPK_optimizer}
\title{A GLPK Optimizer}
\usage{
GLPK_optimizer(control = list(presolve = FALSE))
}
\arguments{
\item{control}{a list of controls}
}
\description{
A GLPK Optimizer
}
